+++ .eslintrc.js
@@ -14,10 +14,6 @@ module.exports = {
       "error",
       "unix"
     ],
-    "quotes": [
-      "error",
-      "single"
-    ],
     "semi": [
       "error",
       "always"