+++ rollup.config.js
@@ -1,7 +1,7 @@
-import uglify from 'rollup-plugin-uglify';
+import uglify from "rollup-plugin-uglify";
 
 const browserPlugins = [];
-if (process.env.NODE_ENV === 'production') {
+if (process.env.NODE_ENV === "production") {
   browserPlugins.push(uglify());
 }
 
@@ -9,22 +9,22 @@ export default [
   // Compressed (for direct consumption in browser)
   // written to dist folder, minified.
   {
-    input: 'src/fecha.js',
+    input: "src/fecha.js",
     plugins: browserPlugins,
     output: {
       // How it will be exposed on window
-      name: 'fecha',
-      format: 'umd',
-      file: 'dist/fecha.min.js',
+      name: "fecha",
+      format: "umd",
+      file: "dist/fecha.min.js"
     }
   },
   // For Node: no minify, output in lib dir
   {
-    input: 'src/fecha.js',
+    input: "src/fecha.js",
     output: {
-      name: 'fecha',
-      format: 'umd',
-      file: 'lib/fecha.umd.js',
+      name: "fecha",
+      format: "umd",
+      file: "lib/fecha.umd.js"
     }
-  },
+  }
 ];