+++ CHANGELOG.md
@@ -1,3 +1,50 @@
+## 4.0.0-beta.0
+** Major Features and  Breaking changes in this version**
+
+#### Improvements
+- *Valid date parsing* - By default fecha will check validity of dates. Previously `2019-55-01` or `2019-01-42` would parse correctly, since Javascript can handle it. Now invalid dates will return `null` instead
+- *ES Module and Tree Shaking Support* - You can now import fecha parseDate or formatDate independently
+
+
+#### Breaking changes
+- `parseDate` may return `null` when previously returned a `Date`. See improvements above, but invalid dates will return `null` now
+- Import and function name changes
+Previously
+```js
+import fecha from 'fecha';
+
+fecha.format(...)
+fecha.parse(...)
+```
+New
+```js
+import {formatDate, parseDate} from 'fecha';
+
+formatDate(...)
+parseDate(...)
+```
+- Change to how to set masks and i18n
+Previously
+```js
+import fecha from 'fecha';
+
+fecha.i18n = { ... }
+fecha.masks.myMask = 'DD , MM, YYYY' 
+```
+
+New
+```js
+import {parseDate, formatDate, setGlobalDateI18n, setGlobalDateMasks} from 'fecha';
+
+setGlobalDateI18n({
+    // ...
+})
+setGlobalDateMasks({
+  myMask: 'DD , MM, YYYY' ;
+});
+```
+
+
 ### 3.0.3
 - Fixed bug when using brackets when parsing dates
 ### 3.0.2